requirements.txt
requests==2.28.2
gitpython==3.1.31
nodeenv==1.6.0
openai==0.12.0
transformers==4.41.2